Hawk/DataScrappingScript/Data_extractor-master/requirements.txt
pymongo==4.6.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0